+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.0.44 Release:
 ## Bug Fixes:
 - Hunter's Mark
+- Grapple
 
 # 1.0.43 Release:
 ## Bug Fixes: